refactor(charts): lift AreaChart + helpers to @skyhook-io/k8s-ui
Move the generic SVG time-series chart primitives out of radar-app into
the shared package, so Radar Hub (and any other downstream library
consumer) can reuse them for custom dashboards without going through the
RadarApp public component.

New `@skyhook-io/k8s-ui/components/charts` subpath exports:
- AreaChart — multi-series SVG chart with hover crosshair, tooltip
  filtering by series sample range, request/limit reference-line overlay
- MetricsSummary — Current / Average / Peak pills (slim interface taking
  a single accent-color class instead of Radar's CategoryDef)
- SeriesLegend — colored dots + short labels with prefix-stripping
- SERIES_COLORS, seriesColor, seriesFill, computeShortLabels — multi-series
  coloring + label-shortening helpers
- formatMetricValue, formatTimestamp — chart-axis formatters
- PrometheusDataPoint, PrometheusSeries, ReferenceLine — types

Radar-side adapters stay in PrometheusCharts.tsx:
- A thin MetricsSummary wrapper that adapts CategoryDef → slim props,
  keeping existing call sites unchanged
- CategoryDef, WORKLOAD_CATEGORIES, NODE_CATEGORIES, TIME_RANGES,
  computeRequestLimitLines stay Radar-specific (Prom-flavored config and
  K8s-resource-derived overlay math)

package.json gains `./components/charts` + `./components/charts/*` export
entries so Hub can deep-import without relying on Vite source aliases.
Backwards-compatible: existing PrometheusCharts.tsx re-exports the lifted
symbols so sibling-file consumers (PrometheusChartsGrid) keep working
either way. PrometheusChartsGrid now imports directly from k8s-ui.

Visually verified on radar-bench: charts render identically, zero
console errors. tsc clean across both packages.
