+++ packages/k8s-ui/src/components/charts/SeriesLegend.tsx
@@ -1,0 +1,30 @@
+import { seriesColor } from './colors'
+import type { PrometheusSeries } from './types'
+
+/**
+ * SeriesLegend — colored dots + pod labels for multi-series charts.
+ * Caps at 10 visible entries (matches AreaChart's distinct color count);
+ * extras collapse to a "+N more" summary.
+ */
+export function SeriesLegend({ series, color }: { series: PrometheusSeries[]; color: string }) {
+  const labels = series.map((s, i) => s.labels.pod || s.labels.instance || `series-${i}`)
+  return (
+    <div className="flex flex-wrap gap-x-4 gap-y-1 px-1">
+      {series.slice(0, 10).map((_, i) => {
+        const shortName = labels[i].length > 40 ? '...' + labels[i].slice(-37) : labels[i]
+        return (
+          <div key={i} className="flex items-center gap-1.5 text-xs text-theme-text-tertiary">
+            <div
+              className="w-2.5 h-2.5 rounded-full shrink-0"
+              style={{ backgroundColor: seriesColor(i, color) }}
+            />
+            <span className="truncate" title={labels[i]}>{shortName}</span>
+          </div>
+        )
+      })}
+      {series.length > 10 && (
+        <span className="text-xs text-theme-text-quaternary">+{series.length - 10} more</span>
+      )}
+    </div>
+  )
+}
